+++ subsys/bluetooth/shell/bt.h
@@ -24,5 +24,6 @@ extern struct bt_conn *default_conn;
 
 int str2bt_addr(const char *str, bt_addr_t *addr);
 void conn_addr_str(struct bt_conn *conn, char *addr, size_t len);
+void hexdump(const struct shell *shell, const u8_t *data, size_t len);
 
 #endif /* __BT_H */
